requirements.txt
requests==2.31.0
pandas==2.1.0
numpy==1.25.0
python-dotenv==1.0.0
